Add ability to list by page type
