+++ directory_cms_client/constants.py
@@ -1,0 +1,8 @@
+EXPORT_READINESS_TERMS_AND_CONDITIONS_SLUG = 'terms-and-conditions'
+EXPORT_READINESS_PRIVACY_AND_COOKIES_SLUG = 'privacy-and-cookies'
+
+FIND_A_SUPPLIER_INDUSTRY_LANDING_SLUG = 'industries-landing-page'
+FIND_A_SUPPLIER_LANDING_SLUG = 'landing-page'
+FIND_A_SUPPLIER_INDUSTRY_CONTACT_SLUG = 'industry-contact'
+FIND_A_SUPPLIER_INDUSTRY_TYPE = 'find_a_supplier.IndustryPage'
+FIND_A_SUPPLIER_INDUSTRY_ARTICLE_TYPE = 'find_a_supplier.IndustryArticlePage'
