+++ directory_cms_client/version.py
@@ -1,1 +1,1 @@
-__version__ = '1.1.0'
+__version__ = '2.0.0'
